clippy: lift suppression of unnecessary_cast (complexity)

+++ source/vir/src/interpreter.rs
@@ -588,7 +588,7 @@ fn u128_to_fixed_width(u: u128, width: u32) -> BigInt {
         16 => BigInt::from_u16(u as u16),
         32 => BigInt::from_u32(u as u32),
         64 => BigInt::from_u64(u as u64),
-        128 => BigInt::from_u128(u as u128),
+        128 => BigInt::from_u128(u),
         _ => panic!("Unexpected fixed-width integer type U({})", width),
     }
     .unwrap()
@@ -601,7 +601,7 @@ fn i128_to_fixed_width(i: i128, width: u32) -> BigInt {
         16 => BigInt::from_i16(i as i16),
         32 => BigInt::from_i32(i as i32),
         64 => BigInt::from_i64(i as i64),
-        128 => BigInt::from_i128(i as i128),
+        128 => BigInt::from_i128(i),
         _ => panic!("Unexpected fixed-width integer type U({})", width),
     }
     .unwrap()
